Initial Release upload
These are the setup and control software files used to create the initial release of the Macchiato Mini-Synth, formally released on February 3. 2017. The control software filename includes the version data v1_161122, indicating the software was finalized and compiled on November 22, 2016. This is the version flashed onto Macchiato units ZD1800-ZD1999.